requirement.txt
# python -m venv env 
# ./env/Scripts/activate

django
djangorestframework
pillow 

#django-admin startproject ____ .

djangorestframework-simplejwt

django-rest-auth 
django-allauth